marchukov/project_02/requirements.txt
kafka
lz4
avro-python3
psycopg2-binary